Create Industrial Revolution
I really like the mixture of two tempos in the beat with the main chords being juxtaposed against what becomes, essentially, a triplet-style synth. I also wanted the beat to feel a bit wider and more spacey, so I panned the hats and reverbed several other sounds. The bitcrusher was just for fun.