Create sound scape
I really wanted to try out using some chord commands so I texted my sister (my secret weapon;) a music major) and she gave me a chord progression that she said was simple
and would sound good so I worked out using cmd-i how I would put in chords and with a little confusion and a lot of time spend I figured it out. I added a little vinyl hiss to soothe the listener into the track. I wanted to see how the computer would improvise with the chords in the background so I used the .choose which sounded pretty cool to me.