Create berlin
My vision here was an artsy nightclub in berlin. The effects are what I really messed around with here, I find it super cool that I could use effects and envelopes to
completely change how things sound. The wobble and envelope on the :blade synth made the sound unique which I really liked. The one that was super hard was the hi hats which came about through using a slicer on a different sound and then realising the rhythm sounded really cool and could be a drum loop.