Create near-death
I wanted to switch it up and make something that would fit into a movie scene. This one is what I imagine you hear after a bomb goes off and the protagonist is in shock. There is nothing to hear but their own heartbeat and the blood flowing through their veins.