Create evil ringtone
This ~beautiful~ jingle is an alarm that I feel would actually wake me up in the morning. It starts off quite similar to mine with some calming noises so it's not too abrupt.
However, a couple of measures in it gets intense but not cluttered so I would actually turn my head and try and get out of bed. The use of the scale helped keep things more streamlined and straightforward rather than trying random numbers/notes until it sounds good.