Create pushing through
The echo tool is what I really messed about with here and what I had fun with. There are so many ways to manipulate a sort of pluck sound with some echo. I did have trouble with the double tick though, they were in the same live loop but different lines and I couldn't get them to work how I wanted them to which is a bit frustrating.