Create truckbed money "money so big"
This feels like it should be my life's work... This is an attempt at a remix of one of my favourite songs called "money so big" by Yeat. I had a digital assignment in my band class when I was unable to play the trumpet (long story) that tasked me with remaking songs I liked. So I have come to enjoy the task quite a bit and doing it in Sonic Pi feels all the more rewarding.